+++ cli/add.c
@@ -6,16 +6,16 @@
 #include "aux/config.h"
 #include "aux/toggle.h"
 
-/* Note: should be static cuz args->task point to it.  */
-static char gentask[IDSIZ + 1];
-
-static int generate_task(tec_arg_t *args, char **argv, int argv_idx)
+static int generate_task(tec_arg_t *args, tec_argvec_t *argvec)
 {
+    char gentask[IDSIZ + 1] = { 0 };
+
     args->taskid = gentask;
     for (register unsigned int i = 1; i < IDLIMIT; ++i) {
         sprintf(gentask, IDFMT, i);
         if (tec_task_exist(teccfg.base.task, args) != LIBTEC_OK) {
-            argv[argv_idx] = gentask;
+            argvec_add(argvec, gentask);
+            args->taskid = argvec->argv[argvec->used];
             return 0;
         }
     }
@@ -41,10 +41,9 @@ static int generate_units(tec_ctx_t *ctx, char *env, char *task)
     return 0;
 }
 
-int tec_cli_add(int argc, const char **argv, tec_ctx_t *ctx)
+int tec_cli_add(tec_argvec_t *argvec, tec_ctx_t *ctx)
 {
     tec_arg_t args;
-    tec_argvec_t argvec;
     const char *errfmt;
     int c, i, retcode, status;
     int opt_quiet, opt_help, opt_cd_dir, opt_cd_toggle;
@@ -55,9 +54,7 @@ int tec_cli_add(int argc, const char **argv, tec_ctx_t *ctx)
     args.env = args.desk = args.taskid = NULL;
     errfmt = "cannot create task '%s': %s";
 
-    argvec_init(&argvec);
-    argvec_parse(&argvec, argc, argv);
-    while ((c = getopt(argvec.used, argvec.argv, ":d:e:hqnN")) != -1) {
+    while ((c = getopt(argvec->used, argvec->argv, ":d:e:hqnN")) != -1) {
         switch (c) {
         case 'd':
             args.desk = optarg;
@@ -93,14 +90,14 @@ int tec_cli_add(int argc, const char **argv, tec_ctx_t *ctx)
         return status;
     else if ((status = check_arg_desk(&args, errfmt, opt_quiet)))
         return status;
-    else if (optind == argc && generate_task(&args, argvec.argv, i)) {
+    else if (optind == argvec->used && generate_task(&args, argvec)) {
         if (opt_quiet == false)
             elog(1, "could not generate task ID: limit is %d", IDLIMIT);
         return 1;
     }
 
     do {
-        args.taskid = argvec.argv[i];
+        args.taskid = argvec->argv[i];
 
         if ((status = tec_task_valid(teccfg.base.task, &args))) {
             if (opt_quiet == false)
@@ -139,11 +136,10 @@ int tec_cli_add(int argc, const char **argv, tec_ctx_t *ctx)
         }
         ctx->units = tec_unit_free(ctx->units);
         retcode = status == LIBTEC_OK ? retcode : status;
-    } while (++i < argvec.used);
+    } while (++i < argvec->used);
 
     if (retcode == LIBTEC_OK && opt_cd_dir)
         retcode = tec_pwd_task(&args) == LIBTEC_OK ? retcode : status;
 
-    argvec_free(&argvec);
     return retcode;
 }