+++ cli/mv.c
@@ -270,10 +270,9 @@ static int parse_dest(const char *path, tec_arg_t *args, int *is_dir,
     return parse_path(path, args, errfmt);
 }
 
-int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
+int tec_cli_mv(tec_argvec_t *argvec, tec_ctx_t *ctx)
 {
     char c;
-    tec_argvec_t argvec;
     int i, showhelp, status, nargs, is_dir;
     tec_arg_t dst, src;
     char *errfmt;
@@ -284,9 +283,7 @@ int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
     src.env = src.desk = src.taskid = NULL;
     dst.env = dst.desk = dst.taskid = NULL;
 
-    argvec_init(&argvec);
-    argvec_parse(&argvec, argc, argv);
-    while ((c = getopt(argvec.used, argvec.argv, ":ft:h")) != -1) {
+    while ((c = getopt(argvec->used, argvec->argv, ":ft:h")) != -1) {
         switch (c) {
         case 'f':
             return elog(1, "option `-f' under development");
@@ -308,7 +305,7 @@ int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
         return help_usage("mv");
 
     i = optind;
-    nargs = argc - i;
+    nargs = argvec->used - i;
 
     if (nargs < 1) {
         return elog(1, "source task is missing");
@@ -319,12 +316,13 @@ int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
     }
 
     /* Parse destination (last argument) */
-    if ((status = parse_dest(argv[argc - 1], &dst, &is_dir, errfmt)))
+    if ((status =
+         parse_dest(argvec->argv[argvec->used - 1], &dst, &is_dir, errfmt)))
         return status;
 
     if (nargs == 2 && !is_dir) {
         /* Single move: tec move src dst */
-        if ((status = parse_path(argv[i], &src, errfmt)))
+        if ((status = parse_path(argvec->argv[i], &src, errfmt)))
             return status;
 
         /* If destination has no env/desk, inherit from source */
@@ -353,11 +351,11 @@ int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
         int last_status = 0;
 
         /* Iterate over all source arguments (all except the last one) */
-        for (; i < argc - 1; i++) {
+        for (; i < argvec->used - 1; i++) {
             /* Reset src for each iteration */
             src.env = src.desk = src.taskid = NULL;
 
-            if ((status = parse_path(argv[i], &src, errfmt))) {
+            if ((status = parse_path(argvec->argv[i], &src, errfmt))) {
                 last_status = status;
                 continue;
             }
@@ -389,7 +387,5 @@ int tec_cli_mv(int argc, const char **argv, tec_ctx_t *ctx)
         }
         return last_status;
     }
-
-    argvec_free(&argvec);
     return 0;
 }