+++ cli/pgn.c
@@ -12,28 +12,25 @@
     3. Separate plugin options from plugin command options.
     4. Or maybe it's better to let the plugin to handle plugin options and the rest.
 */
-int tec_cli_pgn(int argc, const char **argv, tec_ctx_t *ctx)
+int tec_cli_pgn(tec_argvec_t *argvec, tec_ctx_t *ctx)
 {
     int i;
     char *name;
-    tec_argvec_t argvec;
     char cmd[BUFSIZ + 1] = { 0 };
     const char *fmt = "%s/%s/%s -T %s -P %s";
 
     i = 0;
-    argvec_init(&argvec);
-    argvec_parse(&argvec, argc, argv);
-    name = argvec.argv[i++];
+    name = (argvec->argv)[i];
+    argvec_offset(argvec, 1);
 
     sprintf(cmd, fmt, teccfg.base.pgn, name, name, teccfg.base.task,
             teccfg.base.pgn);
 
-    for (; i < argvec.used; ++i) {
+    for (; i < argvec->used; ++i) {
         strcat(cmd, " ");
-        strcat(cmd, argvec.argv[i]);
+        strcat(cmd, (argvec->argv)[i]);
     }
 
     dlog(1, "pgn: %s", cmd);
-    argvec_free(&argvec);
     return system(cmd);
 }