+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `cli/ls`: Remove unnecessary hotfix
 - `cli/add`: Fix return code in case of invalid task ID
 - `cli/env`: Fix memory leak
+- `cli`: Use dynamic array of arguments for API functions
 
 ### Deleted